requirements.txt
selenium==4.29.0
python-dotenv==1.0.1